requirements-worker.txt
polars==1.0.0
pyarrow==15.0.0
pymongo==4.6.1
azure-storage-blob==12.19.0
boto3==1.34.34
mysql-connector-python==8.3.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.10.5
pydantic-settings==2.7.1
numpy<2
